UNDO: capture.h autoformat
