--- include/cv_camera/capture.hpp
@@ -1,278 +1,0 @@
-// Copyright [2015] Takashi Ogura<t.ogura@gmail.com>
-
-#ifndef CV_CAMERA_CAPTURE_H
-#define CV_CAMERA_CAPTURE_H
-
-#include <chrono>
-#include <string>
-#include <thread>
-#include "cv_camera/exception.h"
-
-#include <cv_bridge/cv_bridge.h>
-#include <camera_info_manager/camera_info_manager.hpp>
-#include <image_transport/image_transport.hpp>
-#include <opencv2/highgui/highgui.hpp>
-#include <rclcpp/rclcpp.hpp>
-#include <sensor_msgs/image_encodings.hpp>
-#include <sensor_msgs/msg/camera_info.hpp>
-#include "std_msgs/msg/u_int8.hpp"
-
-// Custom libraries
-#include "utils/console.hpp"
-
-/**
- * @brief namespace of this package
- */
-namespace cv_camera {
-
-/**
- * @brief captures by cv::VideoCapture and publishes to ROS topic.
- *
- */
-class Capture
-{
-   public:
-    /**
-     * @brief costruct with ros node and topic settings
-     *
-     * @param node ROS node handle for advertise topic.
-     * @param topic_name name of topic to publish (this may be image_raw).
-     * @param buffer_size size of publisher buffer.
-     * @param frame_id frame_id of publishing messages.
-     */
-    Capture(rclcpp::Node::SharedPtr node, const std::string& topic_name, uint32_t buffer_size,
-            const std::string& frame_id);
-
-    /**
-     * @brief Open capture device with device ID.
-     *
-     * @param device_id id of camera device (number from 0)
-     * @throw cv_camera::DeviceError device open failed
-     *
-     */
-    bool open(int32_t device_id);
-
-    /**
-     * @brief Open capture device with device name.
-     *
-     * @param port path of the camera device
-     * @throw cv_camera::DeviceError device open failed
-     */
-    bool open(const std::string& port);
-
-    /**
-     * @brief Finds the equivalent camera device associated to
-     *        one port.
-     * @param command Terminal command to be executed
-     * @return command_output::Command output delivered by the terminal
-     */
-    std::string execute_command(const char* command);
-
-    /**
-     * @brief Finds the equivalent camera device associated to
-     *        one port.
-     * @param port Port number of interest
-     * @return port::DeviceError device open failed
-     */
-    std::string det_device_path(const char* port);
-
-    /**
-     * @brief Load camera info from file.
-     *
-     * This loads the camera info from the file specified in the camera_info_url parameter.
-     */
-    void loadCameraInfo();
-
-    /**
-     * @brief Open default camera device.
-     *
-     * This opens with device 0.
-     *
-     * @throw cv_camera::DeviceError device open failed
-     */
-    void open();
-
-    /**
-     * @brief Close capture device.
-        Uses release OpenCV function.
-     */
-    void close();
-
-    /**
-     * @brief open video file instead of capture device.
-     */
-    bool openFile(const std::string& file_path);
-
-    /**
-     * @brief capture an image and store.
-     *
-     * to publish the captured image, call publish();
-     * @return true if success to capture, false if not captured.
-     */
-    bool capture();
-
-    /**
-     * @brief pull an image from the camera but dont decode it
-     *
-     * @return true if success to pull, false if not pulled.
-     */
-    bool grab();
-
-    /**
-     * @brief Publish the image that is already captured by capture().
-     *
-     */
-    void publish();
-
-    /**
-     * @brief accessor of CameraInfo.
-     *
-     * you have to call capture() before call this.
-     *
-     * @return CameraInfo
-     */
-    inline const sensor_msgs::msg::CameraInfo& getInfo() const { return info_; }
-
-    /**
-     * @brief accessor of cv::Mat
-     *
-     * you have to call capture() before call this.
-     *
-     * @return captured cv::Mat
-     */
-    inline const cv::Mat& getCvImage() const { return bridge_.image; }
-
-    /**
-     * @brief accessor of ROS Image message.
-     *
-     * you have to call capture() before call this.
-     *
-     * @return message pointer.
-     */
-    inline const sensor_msgs::msg::Image::SharedPtr getImageMsgPtr() const { return bridge_.toImageMsg(); }
-
-    /**
-     * @brief try capture image width
-     * @return true if success
-     */
-    inline bool setWidth(int32_t width) { return cap_.set(cv::CAP_PROP_FRAME_WIDTH, width); }
-
-    /**
-     * @brief try capture image height
-     * @return true if success
-     */
-    inline bool setHeight(int32_t height) { return cap_.set(cv::CAP_PROP_FRAME_HEIGHT, height); }
-
-    /**
-     * @brief set CV_PROP_*
-     * @return true if success
-     */
-    bool setPropertyFromParam(int property_id, const std::string& param_name);
-
-   private:
-    /**
-     * @brief rescale camera calibration to another resolution
-     */
-    void rescaleCameraInfo(uint width, uint height);
-
-    std::string mat_type2encoding(int mat_type)
-    {
-        switch (mat_type)
-        {
-            case CV_8UC1:
-                return "mono8";
-            case CV_8UC3:
-                return "bgr8";
-            case CV_16SC1:
-                return "mono16";
-            case CV_8UC4:
-                return "rgba8";
-            default:
-                throw std::runtime_error("Unsupported encoding type");
-        }
-    }
-
-    void set_now(builtin_interfaces::msg::Time& time)
-    {
-        std::chrono::nanoseconds now = std::chrono::high_resolution_clock::now().time_since_epoch();
-        if (now <= std::chrono::nanoseconds(0))
-        {
-            time.sec = time.nanosec = 0;
-        }
-        else
-        {
-            time.sec = static_cast<builtin_interfaces::msg::Time::_sec_type>(now.count() / 1000000000);
-            time.nanosec = now.count() % 1000000000;
-        }
-    }
-
-    /**
-     * @brief node handle for advertise.
-     */
-    rclcpp::Node::SharedPtr node_;
-
-    /**
-     * @brief ROS image transport utility.
-     */
-    image_transport::ImageTransport it_;
-
-    /**
-     * @brief name of topic without namespace (usually "image_raw").
-     */
-    std::string topic_name_;
-
-    /**
-     * @brief header.frame_id for publishing images.
-     */
-    std::string frame_id_;
-    /**
-     * @brief size of publisher buffer
-     */
-    uint32_t buffer_size_;
-
-    /**
-     * @brief image publisher created by image_transport::ImageTransport.
-     */
-    image_transport::CameraPublisher pub_;
-
-    /**
-     * @brief capture device.
-     */
-    cv::VideoCapture cap_;
-
-    /**
-     * @brief this stores last captured image.
-     */
-    cv_bridge::CvImage bridge_;
-
-    /**
-     * @brief this stores last captured image info.
-     *
-     * currently this has image size (width/height) only.
-     */
-    sensor_msgs::msg::CameraInfo info_;
-
-    /**
-     * @brief camera info manager
-     */
-    camera_info_manager::CameraInfoManager info_manager_;
-
-    /**
-     * @brief Final publisher for image messages
-     */
-    rclcpp::Publisher<sensor_msgs::msg::Image>::SharedPtr m_pub_image_ptr;
-
-    /**
-     * @brief rescale_camera_info param value
-     */
-    bool rescale_camera_info_;
-
-    /**
-     * @brief capture_delay param value
-     */
-    rclcpp::Duration capture_delay_;
-};
-
-}  // namespace cv_camera
-
-#endif  // CV_CAMERA_CAPTURE_H
